Move tagExpressionSignatureStatus to utils
