+++ executor/src/core/Perspective.ts
@@ -9,8 +9,7 @@ import { MainConfig } from "./Config";
 import { Mutex } from 'async-mutex'
 import { DID } from "@coasys/ad4m/lib/src/DID";
 import { Ad4mDb } from "./db";
-import { getPubSub } from "./utils";
-import Signatures from "./agent/Signatures";
+import { getPubSub, tagExpressionSignatureStatus } from "./utils";
 
 const maxRetries = 10;
 const backoffStep = 200;
@@ -694,7 +693,7 @@ export default class Perspective {
 
         for(const link of values) {
             try {
-                await Signatures.tagExpressionSignatureStatus(link)
+                await tagExpressionSignatureStatus(link)
             } catch (e) {
                 console.error(`Perspective.getLinks(): NH [${this.sharedUrl}] (${this.name}): Got error when trying to tag expression signature status: ${e}`);
                 link.proof.invalid = true;
